requirements.txt
Click==7.0
Django==2.2
django-filter==2.1.0
django-rq==2.0
djangorestframework==3.9.2
gunicorn==19.9.0
Markdown==3.1
pytz==2019.1
redis==3.2.1
rq==1.0
sqlparse==0.3.0
